+++ src/main/kotlin/dev/aaronhowser/mods/irregular_implements/item/DiviningRodItem.kt
@@ -36,6 +36,8 @@ class DiviningRodItem(properties: Properties) : Item(properties) {
 		}
 
 		tooltipComponents.add(component)
+
+		ModTooltipLang.addSodiumTooltip(tooltipComponents)
 	}
 
 	companion object {